+++ apps/headless-crawler/infra/constructs/jobDetailExtractor.ts
@@ -17,7 +17,7 @@ export class JobDetailExtractorConstruct extends Construct {
   ) {
     super(scope, id);
     this.extractor = new NodejsFunction(this, "ScrapingFunction", {
-      entry: "functions/extractTransformAndSaveJobDetailHandler/handler.ts",
+      entry: "functions/extractJobDetailHandler/handler.ts",
       handler: "handler",
       runtime: lambda.Runtime.NODEJS_22_X,
       memorySize: 1024,
