+++ apps/headless-crawler/lib/jobDetailExtractor/index.ts
